requirements.txt
Flask
sentence-transformers
PyPDF2  # O pypdf
faiss-cpu